resource/example_gui/requirements.txt
pyserial>=3.5,<4
PySide6>=6.6,<7
pyqtgraph>=0.13,<0.14
numpy>=1.24,<3
pyinstaller>=6,<7
pyinstaller-hooks-contrib
